Rearrange post learning assessment service tests
